+++ example/android/app/src/main/kotlin/dev/ntfnd404/ephemeral_bloc_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package dev.ntfnd404.ephemeral_bloc_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()