+++ test/fakes/spy_ephemeral_bloc_observer.dart
@@ -4,11 +4,10 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 /// Spy [BlocObserver] that records every [onAction] call for later assertion.
 final class SpyEphemeralBlocObserver extends BlocObserver
     with EphemeralBlocObserver {
-  final List<({BlocBase<Object?> bloc, EphemeralBlocChange<Object?> change})>
-  records = [];
+  final List<({BlocBase<Object?> bloc, Object action})> records = [];
 
   @override
-  void onAction(BlocBase<Object?> bloc, EphemeralBlocChange<Object?> change) {
-    records.add((bloc: bloc, change: change));
+  void onAction(BlocBase<Object?> bloc, Object action) {
+    records.add((bloc: bloc, action: action));
   }
 }